src/crawlers/requirements.txt
dotenv
aiohttp
azure-cosmos
azure-storage-blob
backoff
openai
feedparser
psycopg2-binary
markitdown